Convert ScalarConst Null filter to ConstTableGet with False Filter
create table t1 (a bool);
explain select * from t1 where a = NULL::bool;

We can convert this to a Const Table Get with false filter.

When we have such a scalar, the max card of the operator will be 0, which
will cause it to be converted to a ConstTableGet with False filter.