IndexNLJ should not broadcast co-located tables
Say that we have the following tables:
foo distributed by (a,b)
bar distributed by (c,d) with an index on d

For the query
select * from foo join bar on a = c and b = d;

It will generate a PhysicalFilter on top of an IndexGet on bar. In this
situation, neither the PhysicalFilter nor the IndexGet can create a
matching distribution spec (a,b) to the outer side in order to generate
a plan using a index NLJ.

Thus, this commit allows the IndexGet to create an intermediary
distribution spec, where it matches only the columns that it can. Then
at the PhysicalFilter, the filter will complete the intermediary spec
with its knowledge.

This also removes dead code that contained checks for ORCA issue #309,
which is now resolved.

Also bump ORCA version to 2.61.0.