Penalize HashJoin with a skewed Redistribute as its child
We determine the skew of Redistribute by calculating its NDV. If the NDV < num
of segments, we decide there is a skew. We then penalize the local cost of
HashJoin with a skew ratio = (num of segs)/ndv.
We put a max limit on penalizing skewed hasjoins because penalizing it too much
will lead to picking Gather motions over redistribute. Which is not a good
thing always.