Fix Direct Dispatch with Agg havings no grouping columns
Currently, if there is no grouping column on the Agg, we return a NULL
constraint from CLogicalGbAgg::PpcDeriveConstraint.

However, if there is no grouping column on the agg, we can simply pass
the constraint from the child. This allows us to attach direct dispatch
information if possible using the constraint available on the child.

For example:
```sql
create table foo (a int);
explain analyze select count(*) from foo where a = 1;;
INFO:  (slice 1) Dispatch command to SINGLE content
                                                          QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=0.00..431.00 rows=1 width=8) (actual time=1.205..1.206 rows=1 loops=1)
   ->  Gather Motion 1:1  (slice1; segments: 1)  (cost=0.00..431.00 rows=1 width=1) (actual time=1.192..1.192 rows=0 loops=1)
         ->  Seq Scan on foo  (cost=0.00..431.00 rows=1 width=1) (never executed)
               Filter: (a = 1)
 Planning time: 22.794 ms
   (slice0)    Executor memory: 59K bytes.
   (slice1)    Executor memory: 42K bytes (seg1).
 Memory used:  128000kB
 Optimizer: Pivotal Optimizer (GPORCA) version 3.67.0
 Execution time: 2.506 ms
```

In `CTranslatorExprToDXLUtils::SetDirectDispatchInfo` if there is no contraint, we don't
attach the direct dispatch info.
```
	CPropConstraint *ppc = pdpRel->Ppc();
	if (NULL == ppc->Pcnstr())
	{
		return;
	}
```
For a node, the bare minimum constraint can be the constraint of the child, if
the node itself has no constraint.