Fix stats calculation for predicates with UUID columns
For predicates that involved columns of UUID types, ORCA would consider the
predicate as unsupported and would estimate the selectivity to be 40%. This is
because ORCA did not have statistics for UUID columns. This would cause a bad
plan to be selected in some cases.  This patch fixes the issue by adding
support for the UUID type in CMDIdGPDB class and make it a LINT value for
comparison purposes.