Improve index selection for bitmap indexes
This commit contains the following components
- Prevent btree indexes on AO tables from following the btree path
- Allow col op col predicates as a supported index predicate
- If the predicate is not supported, break it down and combine the
  supported conjuncts instead of keeping them separate
- Use predicate selectivity to determine the best index

Btree indexes on AO tables are not yet supported. Now that such indexes
are sent over to ORCA as type btree instead of type bitmap, we need to
prevent this path from being followed. Instead, btree indexes on AO
tables can be handled by creating a Bitmap index with an underlying
btree index.

In the bitmap index path, alternatives that should not have been
generated (such as picking an index when the first column of the index
was not in the predicate) were being added to the memo because col op
col predicates were being handled separately from col op const
predicates.

In reality, these two scenarios should be following the same code path
when generating viable indexes.

Previously, if the predicate was not all conjuncts, it would be
immediatly broken down into separate predicates. For example, say you
had the predicate a = 4 AND b = 3 AND c = 3 OR (d = 2 AND e = 1), this
would get broken down and ORCA would find indexes for a  = 4, b = 3, c =
3 as individual components.Instead, we want to combine these 3 conjuncts
and find the best matching index.

Additionally, use more than just the number of residuals as an indicator
for a good index. Use the selectivity of the index predicate as an
indiator. If we do not have the NDV of the col, the selectivity
calculation of sel(col = outer ref) = default selectivity of 0.4.

Pick best index based on 1) selectivity, 2) number of residual
predicates and 3) number of columns in the index.

For example, if the predicate was a = 4 and c = 3 and the two indexes
were idx_abc and idx_ac. Say the selectivity was the same, as the number
of residuals is both 0, we want to select the index with the least
number of columns, idx_ac.

Co-Authored-By: Ashuka Xue <axue@pivotal.io>
Co-Authored-By: Hans Zeller <hzeller@pivotal.io>