Fix greedy algorithm to not fallback when it cannot pick a best join component
The greedy algorithm tries to pick the relation from `candidate_component_set`
which gives the minimum cardinality when joined with the existing join tree.
The `candidate_component_set` consists of all the tables which have a
predicated between itself and any table in the current join tree. In certain
cases, it might not be possible to pick a table from the
`candidate_component_set` and in such cases, the xform raises an error and
falls back. For e.g when the candidate component set contains only the right
child of a LOJ. Instead of falling back, the algorithm should try to consider
cross joins and see if it can continue building the join tree. Also if it
cannot construct the join tree, instead of falling back it should not add any
expression to the list of xform results.

Co-authored-by: Shreedhar Hardikar <shardikar@pivotal.io>
Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>