Add new nth-root method to calculate join damping factor
The join damping factor controls how much impact successive predicates
have on the cardinality of a join. For example, given the join condition
"a=b and c=d and e=f", we currently sort the predicates descending by
selectivity and the 2nd predicate does not have as much influence as the
first, depending on the `optimizer_join_factor` GUC. By default, this is
set to .01, meaning the 2nd predicate's contribution to the cardinality
is multiplied by (.01)^2, and the 3rd predicate is multiplied by (.01)^3.

This effectively assumes that predicates are fairly highly correlated.
Although increasing the default value of this GUC was considered (which
would make the assumption that the columns are a bit more independent),
another approach is to scale the successive predicates down moderately
based on their selectivity.

This new method (I'll call it the nth root method) makes successive
predicates have a smaller and smaller impact to account for the higher
probability that there is some correlation between the columns.

For example, given ANDed predicates with selectivities [.5, .3, .1], the
cumulative selectivity would be:
S = S1 * sqrt(S2) * 4root(S3)
.15 = .5 * sqrt(.3) * 4root(.1)
For scale factors, this is equivalent to SF1 * sqrt(SF2) * 4root(SF3)

Using this nth-root method causes us to estimate much lower
cardinalities than before for joins with multiple predicates, causing
plans to use more indexes. This method will not be used by default and
will only be used when the `optimizer_join_factor` GUC is set to 0.

Co-authored-by: Chris Hajas <chajas@pivotal.io>
Co-authored-by: Ashuka Xue <axue@pivotal.io>