Only normalize histogram if well defined
When calculating the statistics of a filter node, the histograms of
newly projected columns are set to empty. Such histograms are not "well
defined" and thus should not be used to derive cardinality. Instead, we
use the default cardinality in such cases.

However, there was a bug which calculated the cardinality from non "well
defined" histograms in the presence of a disjunction on newly projected
columns. This would result in gross underestimation of expected rows of
the filter.

This commit fixes this issue.

Co-authored-by: Ashuka Xue <axue@pivotal.io>
Co-authored-by: Shreedhar Hardikar <shardikar@pivotal.io>