Execute on segment Function Scan JOIN External Scan
Following query gave unexpected results:
  ```
  EXPLAIN SELECT * FROM gp_toolkit.__gp_localid, gp_toolkit.__gp_param_local_setting('gp_role');
                         QUERY PLAN
  -------------------------------------------------------------
   Nested Loop
     Join Filter: true
     ->  Function Scan on __gp_param_local_setting
     ->  Materialize
           ->  Gather Motion 3:1
                 ->  External Scan on __gp_localid
   Optimizer status: PQO version 3.122.0
  ```

Notice that the Function Scan runs on corredinator node. However, the
expectation is to run on the segment. After this patch we generate the
following plan:
  ```
                         QUERY PLAN
  -------------------------------------------------------------
   Gather Motion 3:1
     ->  Nested Loop
           Join Filter: true
           ->  External Scan on __gp_localid
           ->  Function Scan on __gp_param_local_setting
   Optimizer status: PQO version 3.122.0
  ```

Let's consider following tree during optimization (r=requires, d=derives):
  ```
                   InnerNLJoin
                    /       \        .
   r:ANY           /         \   r:SINGLETON
   d:UNIVERSAL    /           \  d:EXTERNAL
                 /             \     .
         FunctionScan        ExternalScan
  ```

On the outer side UNIVERSAL satisifies ANY, thus no enforcer is needed.
However, on the inner side EXTERNAL doesn't satisfy SINGLETON. But we
need to carefully consider how to implement the enforcer. If we apply a
GATHER MOTION, then the function scan may end up executing on the
coordinator node. Instead, we apply a BROADCAST MOTION to ensure that
the function scan executes on the segment nodes.